--- src/app/styles/component/header.scss
@@ -1,37 +1,0 @@
-.header {
-    align-items: center;
-    background-color: var(--color-primary-background);
-    border-bottom: 2px solid var(--color-primary-background-light);;
-    display: flex;
-    justify-content: space-between;
-    padding: 1rem 2rem 0.5rem;
-
-    &__logo {
-        @include accessibility-focus();
-        align-items: center;
-        display: flex;
-        text-decoration: none;
-
-        &-asset {
-            height: 40px;
-            fill: var(--color-primary-text);
-            margin-right: 0.75rem;
-            width: 40px;
-        }
-
-        &-text {
-            color: var(--color-primary-text);
-            font-family: var(--font-secondary);
-            font-size: 2.4rem;
-        }
-    }
-
-    &__controls {
-        display: inline-flex;
-        font-size: 1.6rem;
-
-        p {
-            margin: 0 1rem 0 0;
-        }
-    }
-}
