--- src/app/styles/component/list/list-buttons.scss
@@ -1,42 +1,0 @@
-.list-buttons {
-    display: flex;
-    list-style: none;
-    margin: 2rem 0;
-    padding: 0;
-
-    @include media-breakpoint-down(lg) {
-        flex-direction: column;
-    }
-    
-    > li {
-        padding: 0;
-
-        @include media-breakpoint-up(lg) {
-            margin: 0 1rem;
-        }
-
-        @include media-breakpoint-down(lg) {
-            margin: 1rem 0;
-        }
-
-        &:first-child {
-            @include media-breakpoint-up(lg) {
-                margin-left: 0;
-            }
-
-            @include media-breakpoint-down(lg) {
-                margin-top: 0;
-            }
-        }
-
-        &:last-child {
-            @include media-breakpoint-up(lg) {
-                margin-right: 0;
-            }
-
-            @include media-breakpoint-down(lg) {
-                margin-bottom: 0;
-            }
-        }
-    }
-}
