--- src/app/styles/component/loading.scss
@@ -1,30 +1,0 @@
-.loading {
-    pointer-events: none;
-    position: relative;
-
-    &:before,
-    &:after {
-        bottom: 0;
-        content: "";
-        display: block;
-        height: 100%;
-        left: 0;
-        position: absolute;
-        right: 0;
-        top: 0;
-        width: 100%;
-    }
-
-    &:before {
-        background-color: rgba(0, 0, 0, 0.5);
-        z-index: 1;
-    }
-
-    &:after {
-        background-image: url(#{$path-image}/loading.svg);
-        background-position: center center;
-        background-repeat: no-repeat;
-        background-size: 8rem 8rem;
-        z-index: 2;
-    }
-}
