--- src/app/styles/component/section.scss
@@ -1,3 +1,0 @@
-section {
-    margin: 5rem 0;
-}
