--- src/app/styles/config/accessibility/accessibility-focus.scss
@@ -1,10 +1,0 @@
-@mixin accessibility-focus() {
-    outline-color: var(--color-highlight);
-    outline-offset: 0.3rem;
-    outline-style: solid;
-    outline-width: 2px;
-
-    &:not(:focus-visible) {
-        outline: none;
-    }
-}
