--- src/app/styles/config/paths.scss
@@ -1,1 +1,0 @@
-$path-image: '/image';
