--- src/app/styles/global/theme.scss
@@ -1,4 +1,0 @@
-body {
-    background-color: var(--color-primary-background);
-    color: var(--color-primary-text)
-}
