--- src/app/styles/global/typography.scss
@@ -1,49 +1,0 @@
-html {
-    font-size: 10px;
-}
-
-body {
-    color: var(--color-secondary-text);
-    font-family: var(--font-primary);
-}
-
-h1 {
-    color: var(--color-primary-text);
-    font-family: var(--font-secondary);
-    font-size: 4rem;
-    font-weight: 700;
-    margin: 0 0 1rem 0;
-}
-
-h2 {
-    color: var(--color-primary-text);
-    font-family: var(--font-secondary);
-    font-size: 3rem;
-    font-weight: 700;
-    margin: 1.5rem 0 1.5rem 0;
-}
-
-p {
-    margin: 0 0 2rem 0;
-}
-
-hr {
-   border-top: 1px solid var(--color-primary-text);
-   margin: 5rem 0;
-}
-
-ul, ol {
-    margin: 0 0 2.5rem;
-
-    li {
-        margin: 0.4rem 0;
-    }
-}
-
-a {
-    text-decoration: underline;
-
-    &:hover {
-        text-decoration: none;
-    }
-}
