+++ src/styles/component/button/button-solid.scss
@@ -1,0 +1,36 @@
+.button-solid {
+    @include accessibility-focus();
+    align-items: center;
+    background-color: var(--color-primary-background-light);
+    border: 1px solid var(--color-primary-background-light);
+    border-radius: 0.5rem;
+    color: var(--color-primary-text);
+    display: inline-flex;
+    font-family: var(--font-secondary);
+    font-size: 2rem;
+    padding: 0.25rem 0.25rem;
+    text-decoration: none;
+
+    &__text {
+        margin: 0 0.5rem;
+    }
+
+    &__icon {
+        background-repeat: no-repeat;
+        background-size: 100% 100%;
+        height: 2rem;
+        fill: var(--color-primary-text);
+        margin: 0 0.75rem;
+        width: 2rem;
+    }
+
+    // States
+    &:hover {
+        border-color: var(--color-highlight);
+    }
+
+    &[disabled] {
+        opacity: 0.5;
+        pointer-events: none;
+    }
+}
