+++ src/styles/component/footer.scss
@@ -1,0 +1,11 @@
+.footer {
+    background-color: var(--color-primary-background-tinted);
+    border-top: 3px solid var(--color-primary-background-light);
+    font-size: 1.6rem;
+    padding: 3rem 0;
+    text-align: center;
+
+    p {
+        margin: 0;
+    }
+}
