+++ src/styles/component/loading.scss
@@ -1,0 +1,30 @@
+.loading {
+    pointer-events: none;
+    position: relative;
+
+    &:before,
+    &:after {
+        bottom: 0;
+        content: "";
+        display: block;
+        height: 100%;
+        left: 0;
+        position: absolute;
+        right: 0;
+        top: 0;
+        width: 100%;
+    }
+
+    &:before {
+        background-color: rgba(255, 255, 255, 0.75);
+        z-index: 1;
+    }
+
+    &:after {
+        background-image: url(#{$path-image}/loading.svg);
+        background-position: center center;
+        background-repeat: no-repeat;
+        background-size: 8rem 8rem;
+        z-index: 2;
+    }
+}
