+++ src/styles/component/section.scss
@@ -1,0 +1,3 @@
+section {
+    margin: 5rem 0;
+}
