+++ src/styles/config/paths.scss
@@ -1,0 +1,1 @@
+$path-image: '/image';
