+++ src/styles/global/theme.scss
@@ -1,0 +1,4 @@
+body {
+    background-color: var(--color-primary-background);
+    color: var(--color-primary-text)
+}
