+++ src/styles/main.scss
@@ -1,0 +1,18 @@
+@import "library/bootstrap";
+
+@import "config/paths";
+@import "config/accessibility/accessibility-focus";
+@import "config/variables";
+
+@import "global/main-content";
+@import "global/theme";
+@import "global/typography";
+
+@import "component/button/button-solid";
+@import "component/footer";
+@import "component/form/toggle";
+@import "component/header";
+@import "component/hero";
+@import "component/list/list-buttons";
+@import "component/loading";
+@import "component/section";
